Record cost breakdown for RTC and RTD.
There seems to be an issue cased by the difference between RTC and RTD. RTC covers 12 5-min intervals while RTD only covers one. While the load used in the 118 bus system simulation is the same, the solution is totally different. Since no look-ahead intervals in the RTD, the RTD will use a "greedy" strategy to minimize production cost in the only covered interval, at the cost of reserve shortage in the next interval. I don't have a quick solution for this, maybe we should increase the length of look-ahead intervals.